minor changes for better cross-platform support. now include zstd using fetchcontent

+++ zvcr_lib/src/zvcr/common/data_storage.hpp
@@ -179,7 +179,7 @@ namespace zvcr {
 
         [[nodiscard]]
         static uint64_t getBitsPerIndex(const size_t length) {
-            return std::max<uint64_t>(std::bit_width(std::max(length, 1UL) - 1), 1UL);
+            return std::max<uint64_t>(std::bit_width(std::max(static_cast<uint64_t>(length), static_cast<uint64_t>(1)) - 1), 1UL);
         }
 
         [[nodiscard]]
